requirements.txt
oauthlib==0.7.2
requests==2.5.1
requests-oauthlib==0.4.2
tapioca-wrapper==0.2.1
wsgiref==0.1.2
